Reuse code in memory-based io plugins ##refactor
