+++ librz/io/meson.build
@@ -2,6 +2,7 @@ rz_io_sources = [
   'io.c',
   'io_fd.c',
   'io_map.c',
+  'io_memory.c',
   'io_cache.c',
   'io_desc.c',
   'io_plugin.c',
