+++ librz/io/p/io_malloc.c
@@ -1,157 +1,9 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 
-#include "rz_io.h"
 #include "rz_lib.h"
 #include <stdio.h>
 #include <stdlib.h>
-#include <sys/types.h>
-
-typedef struct {
-	ut8 *buf;
-	ut32 size;
-	ut64 offset;
-} RzIOMalloc;
-
-static inline ut32 _io_malloc_sz(RzIODesc *desc) {
-	if (!desc) {
-		return 0;
-	}
-	RzIOMalloc *mal = (RzIOMalloc*)desc->data;
-	return mal? mal->size: 0;
-}
-
-static inline void _io_malloc_set_sz(RzIODesc *desc, ut32 sz) {
-	if (!desc) {
-		return;
-	}
-	RzIOMalloc *mal = (RzIOMalloc*)desc->data;
-	if (mal) {
-		mal->size = sz;
-	}
-}
-
-static inline ut8* _io_malloc_buf(RzIODesc *desc) {
-	if (!desc) {
-		return NULL;
-	}
-	RzIOMalloc *mal = (RzIOMalloc*)desc->data;
-	return mal->buf;
-}
-
-
-static inline ut8* _io_malloc_set_buf(RzIODesc *desc, ut8* buf) {
-	if (!desc) {
-		return NULL;
-	}
-	RzIOMalloc *mal = (RzIOMalloc*)desc->data;
-	return mal->buf = buf;
-}
-
-static inline ut64 _io_malloc_off(RzIODesc *desc) {
-	if (!desc) {
-		return 0;
-	}
-	RzIOMalloc *mal = (RzIOMalloc*)desc->data;
-	return mal->offset;
-}
-
-static inline void _io_malloc_set_off(RzIODesc *desc, ut64 off) {
-	if (!desc) {
-		return;
-	}
-	RzIOMalloc *mal = (RzIOMalloc*)desc->data;
-	mal->offset = off;
-}
-
-static int __write(RzIO *io, RzIODesc *fd, const ut8 *buf, int count) {
-	if (!fd || !buf || count < 0 || !fd->data) {
-		return -1;
-	}
-	if (_io_malloc_off (fd) > _io_malloc_sz (fd)) {
-		return -1;
-	}
-	if (_io_malloc_off (fd) + count > _io_malloc_sz (fd)) {
-		count -= (_io_malloc_off (fd) + count -_io_malloc_sz (fd));
-	}
-	if (count > 0) {
-		memcpy (_io_malloc_buf (fd) + _io_malloc_off (fd), buf, count);
-		_io_malloc_set_off (fd, _io_malloc_off (fd) + count);
-		return count;
-	}
-	return -1;
-}
-
-static bool __resize(RzIO *io, RzIODesc *fd, ut64 count) {
-	ut8 * new_buf = NULL;
-	if (!fd || !fd->data || count == 0) {
-		return false;
-	}
-	ut32 mallocsz = _io_malloc_sz (fd);
-	if (_io_malloc_off (fd) > mallocsz) {
-		return false;
-	}
-	new_buf = malloc (count);
-	if (!new_buf) {
-		return false;
-	}
-	memcpy (new_buf, _io_malloc_buf (fd), RZ_MIN (count, mallocsz));
-	if (count > mallocsz) {
-		memset (new_buf + mallocsz, 0, count - mallocsz);
-	}
-	free (_io_malloc_buf (fd));
-	_io_malloc_set_buf (fd, new_buf);
-	_io_malloc_set_sz (fd, count);
-	return true;
-}
-
-static int __read(RzIO *io, RzIODesc *fd, ut8 *buf, int count) {
-	memset (buf, 0xff, count);
-	if (!fd || !fd->data) {
-		return -1;
-	}
-	ut32 mallocsz = _io_malloc_sz (fd);
-	if (_io_malloc_off (fd) > mallocsz) {
-		return -1;
-	}
-	if (_io_malloc_off (fd) + count >= mallocsz) {
-		count = mallocsz - _io_malloc_off (fd);
-	}
-	memcpy (buf, _io_malloc_buf (fd) + _io_malloc_off (fd), count);
-	_io_malloc_set_off (fd, _io_malloc_off (fd) + count);
-	return count;
-}
-
-static int __close(RzIODesc *fd) {
-	RzIOMalloc *riom;
-	if (!fd || !fd->data) {
-		return -1;
-	}
-	riom = fd->data;
-	RZ_FREE (riom->buf);
-	RZ_FREE (fd->data);
-	return 0;
-}
-
-static ut64 __lseek(RzIO* io, RzIODesc *fd, ut64 offset, int whence) {
-	ut64 rz_offset = offset;
-	if (!fd || !fd->data) {
-		return offset;
-	}
-	ut32 mallocsz = _io_malloc_sz (fd);
-	switch (whence) {
-	case SEEK_SET:
-		rz_offset = (offset <= mallocsz) ? offset : mallocsz;
-		break;
-	case SEEK_CUR:
-		rz_offset = (_io_malloc_off (fd) + offset <= mallocsz ) ? _io_malloc_off (fd) + offset : mallocsz;
-		break;
-	case SEEK_END:
-		rz_offset = _io_malloc_sz (fd);
-		break;
-	}
-	_io_malloc_set_off (fd, rz_offset);
-	return rz_offset;
-}
+#include "../io_memory.h"
 
 static bool __check(RzIO *io, const char *pathname, bool many) {
 	return (!strncmp (pathname, "malloc://", 9)) || (!strncmp (pathname, "hex://", 6));
@@ -160,6 +12,9 @@ static bool __check(RzIO *io, const char *pathname, bool many) {
 static RzIODesc *__open(RzIO *io, const char *pathname, int rw, int mode) {
 	if (__check (io, pathname, 0)) {
 		RzIOMalloc *mal = RZ_NEW0 (RzIOMalloc);
+		if (!mal) {
+			return NULL;
+		}
 		if (!strncmp (pathname, "hex://", 6)) {
 			mal->size = strlen (pathname);
 			mal->buf = calloc (1, mal->size + 1);
@@ -197,12 +52,12 @@ RzIOPlugin rz_io_plugin_malloc = {
 	.uris = "malloc://,hex://",
 	.license = "LGPL3",
 	.open = __open,
-	.close = __close,
-	.read = __read,
+	.close = io_memory_close,
+	.read = io_memory_read,
 	.check = __check,
-	.lseek = __lseek,
-	.write = __write,
-	.resize = __resize,
+	.lseek = io_memory_lseek,
+	.write = io_memory_write,
+	.resize = io_memory_resize,
 };
 
 #ifndef RZ_PLUGIN_INCORE
